Add a faster version of the binomial coefficients
The new implementation uses a recurrence relation to avoid repeatedly
calling the `!` primitive; this results in drastically faster runtime
at the cost of some numerical accuracy, still for J=1000 the largest
relative error compared to the result given by `!` is of the order of
`1e-11`.

I also added some tags to both implementations.